Update readme a bit, pretty up the required imports, remove the QueryNotEquals node, since you can create it from composition with Not and Equals.
