+++ mild-sql-dsl/src/main/scala/scoobie/snacks/mild/sql/sql.scala
@@ -1,14 +1,11 @@
 package scoobie.snacks.mild
 
 import scoobie.ast._
-import scoobie.snacks.mild.sql.primitives._
-import scoobie.snacks.mild.sql.primitives.{QueryComparisonExtensions, QueryProjectionExtensions, QueryValueExtensions, SqlDslStringInterpolators}
+import scoobie.snacks.mild.sql.primitives
 import scoobie.coercion.Coerce
+import scoobie.snacks.TypeComparisons._
 
-/**
- * Created by jacob.barber on 3/4/16.
- */
-package object sql extends query.modify with query.select {
+package object sql extends query.modify with query.select with primitives {
 
   implicit val stringExpr = RawExpressionHandler[String](identity)
 
@@ -43,12 +40,4 @@ package object sql extends query.modify with query.select {
   implicit def toQueryProjection[F[_]](queryPath: QueryPath[F])(implicit coerce: Coerce[F]): QueryProjection[F] =
     QueryProjectOne(queryPath, None)
 
-  /* Stolen From Shapeless */
-  def unexpected : Nothing = sys.error("Unexpected invocation")
-
-  trait =:!=[A, B] extends Serializable
-
-  implicit def neq[A, B] : A =:!= B = new =:!=[A, B] {}
-  implicit def neqAmbig1[A] : A =:!= A = unexpected
-  implicit def neqAmbig2[A] : A =:!= A = unexpected
 }
